requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
idna==3.1
multidict==5.1.0
mysqlclient==2.0.3
neispy==3.2.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
yarl==1.6.3
